feat: added a preliminary content provider, loader support in CommentsActivity, and post options dialog
Given Udacity project requirements, I needed to implement content provider and loader support in the app. Thus, I introduced classes to create SQLite database associated with the app, and a content provider that draws from it. I also added loader support in CommentsActivity. Given the extensive amount of work that would be required to implement Cursor and CursorAdapter at this point, this preliminary application of loader and content provider will likely be the last work on loader and content provider for this app.

I also added a dialog that presents more options when a post is long clicked in MainActivity. This allows for the user to quickly go to a post's subreddit if they're not already in it, or open the post link in their default browser.