fix: started translation of ORM objects to cursor for content provider, and changed widget provider
Introduced methods to translate ContentValues into ORM objects to save in ORM storage, and added translation of ORM objects into cursor to return in query method of content provider. The previously introduced database tables will be abandoned in the future. I also changed the widget provider to enforce a minWidth of 250dp (or 4 cells on the homescreen) to reduce the chances of overflow in certain rows of item_widget. Finally, I added a change in current subreddit preference in SharedPreferences when the user chooses to go to a random subreddit, thus eliminating an issue where rotation or recreation of MainActivity in a random subreddit is accompanied by a name change in the subreddit, but not in the posts.