feat: implemented content provider for sharing currently loaded posts and comments
Implemented query method of content provider to return a cursor according to the requested URI. Supports query for all posts, a specific post, all comments, comments on a specific post and a specific comment. From here, requesting apps can obtain all the information contained in the the requested objects as columns in the cursor.