fix: fixed loader in CommentsActivity to use cursor, CommentObjects are now stored using content provider
For the purpose of the Udacity project, the ORM approach was abandoned for Comments Activity and the RecyclerView Adapter contained within it in favor of storing comment objects in an SQLite database using the Content Provider. Additionally, the loader now returns a cursor loader that is used to load data into the adapter.

Additionally, signing config was added with throwaway password (again for purposes of the Udacity project). This won't be used for the final release version of the app.