fix: signing configuration uses relative path, installRelease task now works
I changed the location of the app_keystore.jks file to allow for it to be found by relative path in the app module. This allows the signing configuration to be found in version control. I also modified installRelease task to directly depend on the respective release-type install tasks to avoid errors that were occurring due to missing dependencies within the tasks that were being found with the previously used task filter.

Finally, I moved the ReaditContract class to each respective product flavor source set to allow for differing authorities for each flavor's ContentProvider, allowing both of them to be installed on a device at the same time. A matching string resource is now referred to in the Manifest to avoid conflicts.