fix: non-author commenters are not marked as post authors, loading view doesn't overlay selftext for previously loaded posts
Previously, non-author comments were being bound as author comments after a single author comment in similar recyclerview positions. This behavior has been fixed by checking for the author background drawable and setting the background drawable to null if the comment's author is not actually the post author.

Previously, when a user navigates to a post's comments, went back to the MainActivity and clicked on the same post again, the animated loading view would overlay the collapsingtoolbar layout until the user scrolls down to the comments, resulting in a misleading loading animation that stays visible especially in posts with a large selftext. This behavior has been fixed by setting the loading view visibility to GONE as soon as the user navigates back to a post whose comments are already loaded into memory.